Consistently naming files and URLS, and removing outdated files
This update renames some of the markdown files to apply a consistent naming
scheme across all files and across all training courses. Specifically,
some files were renamed from singular (Contribution-x) to plural
(Contributions-x).

After splitting module 1 and 3, we kept the full files in case we wanted to
undo the split. As the split went well and we want to keep this structure,
the old files must be removed.

Finally, this change renames the target URL of Contribution-4-Ind to a consistent
Contributions-4-Ind in alignment with all other modules.
Note: This is a breaking change which needs modifications in the PECA player.